Refactor to accommodate both Businesses and Towns under Location as an abstract class.
